# packages/openapi-generator/tmp/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lonestone API",
    "description": "The Lonestone API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "@lonestone",
      "description": ""
    }
  ],
  "servers": [],
  "paths": {
    "/api": {
      "get": {
        "operationId": "AppController_getHello",
        "tags": [
          "App"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/analysis/{id}/analyze": {
      "post": {
        "operationId": "AnalysisController_startAnalyze",
        "tags": [
          "Analysis"
        ],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/analysis/events": {
      "get": {
        "operationId": "AnalysisController_getEvents",
        "tags": [
          "Analysis"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "SSE stream of analysis",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Analysis events schema ": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "step": {
            "type": "string",
            "enum": [
              "extraction",
              "analysis-part-one",
              "analysis-part-two",
              "failed",
              "completed"
            ]
          }
        },
        "required": [
          "id",
          "step"
        ],
        "additionalProperties": false,
        "title": "Analysis events schema ",
        "description": "Analysis events schema"
      }
    }
  }
}